requirements.txt
fastAPI>=0.115.14
uvicorn>=0.34.3
aiomysql>=0.2.0
mysql-connector-python>=9.3.0
python-socketio>=5.13.0